chore: refactor l1 module
